+++ internal/provider/provider.go
@@ -155,8 +155,9 @@ func (p *ProviderFramework) Resources(ctx context.Context) []func() resource.Res
 		// Webhook
 		webhook.NewNotificationsWebhookResource,
 		// Waypoint
+		waypoint.NewActionResource,
 		waypoint.NewApplicationResource,
-		waypoint.NewApplicationTemplateResource,
+		waypoint.NewTemplateResource,
 		waypoint.NewAddOnResource,
 		waypoint.NewAddOnDefinitionResource,
 		waypoint.NewTfcConfigResource,
@@ -179,8 +180,9 @@ func (p *ProviderFramework) DataSources(ctx context.Context) []func() datasource
 		iam.NewGroupDataSource,
 		iam.NewUserPrincipalDataSource,
 		// Waypoint
+		waypoint.NewActionDataSource,
 		waypoint.NewApplicationDataSource,
-		waypoint.NewApplicationTemplateDataSource,
+		waypoint.NewTemplateDataSource,
 		waypoint.NewAddOnDataSource,
 		waypoint.NewAddOnDefinitionDataSource,
 	}, packer.DataSourceSchemaBuilders...)