+++ internal/provider/waypoint/resource_waypoint_application_test.go
@@ -189,7 +189,7 @@ func testAccCheckWaypointApplicationDestroy(t *testing.T, applicationModel *wayp
 // sufficient for now.
 func testApplicationConfig(tempName, appName string) string {
 	return fmt.Sprintf(`
-resource "hcp_waypoint_application_template" "test" {
+resource "hcp_waypoint_template" "test" {
   name    = "%s"
   summary = "some summary for fun"
   readme_markdown_template = base64encode("# Some Readme")
@@ -206,13 +206,13 @@ resource "hcp_waypoint_application_template" "test" {
 
 resource "hcp_waypoint_application" "test" {
   name    = "%s"
-  application_template_id = hcp_waypoint_application_template.test.id
+  template_id = hcp_waypoint_template.test.id
 }`, tempName, appName)
 }
 
 func testApplicationWithInputVarsConfig(tempName, appName string) string {
 	return fmt.Sprintf(`
-resource "hcp_waypoint_application_template" "test_var_opts" {
+resource "hcp_waypoint_template" "test_var_opts" {
   name    = "%s"
   summary = "some summary for fun"
   readme_markdown_template = base64encode("# Some Readme")
@@ -254,7 +254,7 @@ resource "hcp_waypoint_application_template" "test_var_opts" {
 
 resource "hcp_waypoint_application" "test_var_opts" {
   name    = "%s"
-  application_template_id = hcp_waypoint_application_template.test_var_opts.id
+  template_id = hcp_waypoint_template.test_var_opts.id
 
   application_input_variables = [
 	{
@@ -273,7 +273,7 @@ resource "hcp_waypoint_application" "test_var_opts" {
 
 func testApplicationWithNoInputVarsConfig(tempName, appName string) string {
 	return fmt.Sprintf(`
-resource "hcp_waypoint_application_template" "test_var_opts" {
+resource "hcp_waypoint_template" "test_var_opts" {
   name    = "%s"
   summary = "some summary for fun"
   readme_markdown_template = base64encode("# Some Readme")
@@ -308,6 +308,6 @@ resource "hcp_waypoint_application_template" "test_var_opts" {
 
 resource "hcp_waypoint_application" "test_var_opts" {
   name    			      = "%s"
-  application_template_id = hcp_waypoint_application_template.test_var_opts.id
+  template_id = hcp_waypoint_template.test_var_opts.id
 }`, tempName, appName)
 }