+++ internal/provider/waypoint/resource_waypoint_template_test.go
@@ -20,30 +20,30 @@ import (
 	sharedmodels "github.com/hashicorp/hcp-sdk-go/clients/cloud-shared/v1/models"
 )
 
-func TestAccWaypoint_Application_Template_basic(t *testing.T) {
-	var appTemplateModel waypoint.ApplicationTemplateResourceModel
-	resourceName := "hcp_waypoint_application_template.test"
+func TestAccWaypoint_Template_basic(t *testing.T) {
+	var appTemplateModel waypoint.TemplateResourceModel
+	resourceName := "hcp_waypoint_template.test"
 	name := generateRandomName()
 	updatedName := generateRandomName()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
 		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
-		CheckDestroy:             testAccCheckWaypointAppTemplateDestroy(t, &appTemplateModel),
+		CheckDestroy:             testAccCheckWaypointTemplateDestroy(t, &appTemplateModel),
 		Steps: []resource.TestStep{
 			{
-				Config: testAppTemplateConfig(name),
+				Config: testTemplateConfig(name),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckWaypointAppTemplateExists(t, resourceName, &appTemplateModel),
-					testAccCheckWaypointAppTemplateName(t, &appTemplateModel, name),
+					testAccCheckWaypointTemplateExists(t, resourceName, &appTemplateModel),
+					testAccCheckWaypointTemplateName(t, &appTemplateModel, name),
 					resource.TestCheckResourceAttr(resourceName, "name", name),
 				),
 			},
 			{
-				Config: testAppTemplateConfig(updatedName),
+				Config: testTemplateConfig(updatedName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckWaypointAppTemplateExists(t, resourceName, &appTemplateModel),
-					testAccCheckWaypointAppTemplateName(t, &appTemplateModel, updatedName),
+					testAccCheckWaypointTemplateExists(t, resourceName, &appTemplateModel),
+					testAccCheckWaypointTemplateName(t, &appTemplateModel, updatedName),
 					resource.TestCheckResourceAttr(resourceName, "name", updatedName),
 				),
 			},
@@ -51,21 +51,21 @@ func TestAccWaypoint_Application_Template_basic(t *testing.T) {
 	})
 }
 
-func TestAccWaypoint_Application_template_with_variable_options(t *testing.T) {
-	var appTemplateModel waypoint.ApplicationTemplateResourceModel
-	resourceName := "hcp_waypoint_application_template.var_opts_test"
+func TestAccWaypoint_template_with_variable_options(t *testing.T) {
+	var appTemplateModel waypoint.TemplateResourceModel
+	resourceName := "hcp_waypoint_template.var_opts_test"
 	name := generateRandomName()
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { acctest.PreCheck(t) },
 		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
-		CheckDestroy:             testAccCheckWaypointAppTemplateDestroy(t, &appTemplateModel),
+		CheckDestroy:             testAccCheckWaypointTemplateDestroy(t, &appTemplateModel),
 		Steps: []resource.TestStep{
 			{
-				Config: testAppTemplateConfigWithVarOpts(name),
+				Config: testTemplateConfigWithVarOpts(name),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckWaypointAppTemplateExists(t, resourceName, &appTemplateModel),
-					testAccCheckWaypointAppTemplateName(t, &appTemplateModel, name),
+					testAccCheckWaypointTemplateExists(t, resourceName, &appTemplateModel),
+					testAccCheckWaypointTemplateName(t, &appTemplateModel, name),
 					resource.TestCheckResourceAttr(resourceName, "name", name),
 				),
 			},
@@ -73,22 +73,22 @@ func TestAccWaypoint_Application_template_with_variable_options(t *testing.T) {
 	})
 }
 
-// simple attribute check on the application template receved from the API
-func testAccCheckWaypointAppTemplateName(t *testing.T, appTemplateModel *waypoint.ApplicationTemplateResourceModel, nameValue string) resource.TestCheckFunc {
+// simple attribute check on the template receved from the API
+func testAccCheckWaypointTemplateName(t *testing.T, appTemplateModel *waypoint.TemplateResourceModel, nameValue string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		if appTemplateModel.Name.ValueString() != nameValue {
-			return fmt.Errorf("expected application template name to be %s, but got %s", nameValue, appTemplateModel.Name.ValueString())
+			return fmt.Errorf("expected template name to be %s, but got %s", nameValue, appTemplateModel.Name.ValueString())
 		}
 		return nil
 	}
 }
 
-func testAccCheckWaypointAppTemplateExists(t *testing.T, resourceName string, appTemplateModel *waypoint.ApplicationTemplateResourceModel) resource.TestCheckFunc {
+func testAccCheckWaypointTemplateExists(t *testing.T, resourceName string, appTemplateModel *waypoint.TemplateResourceModel) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		// find the corresponding state object
 		rs, ok := s.RootModule().Resources[resourceName]
 		if !ok {
-			return fmt.Errorf("Not found: %s", resourceName)
+			return fmt.Errorf("not found: %s", resourceName)
 		}
 
 		client := acctest.HCPClients(t)
@@ -102,7 +102,7 @@ func testAccCheckWaypointAppTemplateExists(t *testing.T, resourceName string, ap
 			ProjectID:      projectID,
 		}
 
-		// Fetch the application template
+		// Fetch the template
 		template, err := clients.GetApplicationTemplateByID(context.Background(), client, loc, appTempID)
 		if err != nil {
 			return err
@@ -120,7 +120,7 @@ func testAccCheckWaypointAppTemplateExists(t *testing.T, resourceName string, ap
 	}
 }
 
-func testAccCheckWaypointAppTemplateDestroy(t *testing.T, appTemplateModel *waypoint.ApplicationTemplateResourceModel) resource.TestCheckFunc {
+func testAccCheckWaypointTemplateDestroy(t *testing.T, appTemplateModel *waypoint.TemplateResourceModel) resource.TestCheckFunc {
 	return func(_ *terraform.State) error {
 		client := acctest.HCPClients(t)
 		id := appTemplateModel.ID.ValueString()
@@ -144,16 +144,16 @@ func testAccCheckWaypointAppTemplateDestroy(t *testing.T, appTemplateModel *wayp
 		// fall through, we expect a not found above but if we get this far then
 		// the test should fail
 		if template != nil {
-			return fmt.Errorf("expected application template to be destroyed, but it still exists")
+			return fmt.Errorf("expected Template to be destroyed, but it still exists")
 		}
 
 		return fmt.Errorf("both template and error were nil in destroy check, this should not happen")
 	}
 }
 
-func testAppTemplateConfig(name string) string {
+func testTemplateConfig(name string) string {
 	return fmt.Sprintf(`
-resource "hcp_waypoint_application_template" "test" {
+resource "hcp_waypoint_template" "test" {
   name                     = "%s"
   summary                  = "some summary for fun"
   readme_markdown_template = base64encode("# Some Readme")
@@ -169,9 +169,9 @@ resource "hcp_waypoint_application_template" "test" {
 }`, name)
 }
 
-func testAppTemplateConfigWithVarOpts(name string) string {
+func testTemplateConfigWithVarOpts(name string) string {
 	return fmt.Sprintf(`
-resource "hcp_waypoint_application_template" "var_opts_test" {
+resource "hcp_waypoint_template" "var_opts_test" {
   name                     = "%s"
   summary                  = "A template with a variable with options."
   readme_markdown_template = base64encode("# Some Readme")